news_service/requirements.txt
Flask
pymongo
PyJWT
werkzeug
requests
retrying